Added Desert background, removed ExplanationAboutLevels.txt
I feel like the file was useless anyways and if a new level idea was made and added, it would be outdated until it was updated. It was just another stupid thing and now anyone can add a level that doesn't follow the plan and if it's not bad and if it's not too much like what a bonus level would be (like if certain tiles became dangerous or something), then it would be accepted.